testing: rewrite TestRepartitioning as SQLTranslator tests
Fixes #49112

This rewrites TestRepartitioning subtests as SQLTranslator tests so that they
can be unskipped.

Release note: None
