+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/multi_col_list_partitioning_default_to_multi_col_list_partitioning
@@ -1,0 +1,2 @@
+repartition from=../partition/multi_col_list_partitioning to=../partition/multi_col_list_partitioning_default
+----
