+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/multi_col_list_partitioning_to_multi_col_list_partitioning_default
@@ -1,0 +1,2 @@
+repartition from=../partition/multi_col_list_partitioning_default to=../partition/multi_col_list_partitioning
+----
