+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/multi_col_range_partitioning_maxvalue_maxvalue_to_multi_col_range_partitioning_maxvalue
@@ -1,0 +1,2 @@
+repartition from=../partition/multi_col_range_partitioning_maxvalue_maxvalue to=../partition/multi_col_range_partitioning_maxvalue
+----
