+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/secondary_index_list_partitioning_default_to_secondary_index_list_partitioning
@@ -1,0 +1,2 @@
+repartition from=../partition/secondary_index_list_partitioning_default to=../partition/secondary_index_list_partitioning
+----
