+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/secondary_index_list_partitioning_to_secondary_index_list_partitioning_default
@@ -1,0 +1,2 @@
+repartition from=../partition/secondary_index_list_partitioning to=../partition/secondary_index_list_partitioning_default
+----
