+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/secondary_index_unpartitioned_to_secondary_index_list_partitioned
@@ -1,0 +1,2 @@
+repartition from=../partition/secondary_index_unpartitioned to=../partition/secondary_index_list_partitioning
+----
