+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/single_col_list_partitioning_default_to_unpartitioned
@@ -1,0 +1,2 @@
+repartition from=../partition/single_col_list_partitioning_default to=../partition/unpartitioned
+----
