+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/single_col_list_partitioning_to_single_col_list_partitioning_default
@@ -1,0 +1,2 @@
+repartition from=../partition/single_col_list_partitioning to=../partition/single_col_list_partitioning_default
+----
