+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/single_col_range_partitioning_maxvalue_to_unpartitioned
@@ -1,0 +1,2 @@
+repartition from=../partition/single_col_range_partitioning_maxvalue to=../partition/unpartitioned
+----
