+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/unpartitioned_to_single_col_range_partitioning
@@ -1,0 +1,2 @@
+repartition from=../partition/unpartitioned to=../partition/single_col_range_partitioning
+----
