+++ pkg/ccl/spanconfigccl/spanconfigsqltranslatorccl/testdata/repartition/unpartitioned_to_unpartitioned
@@ -1,0 +1,2 @@
+repartition from=../partition/unpartitioned to=../partition/unpartitioned
+----
